+++ src/Components/Facility/ConsultationDialysisTab.tsx
@@ -5,9 +5,8 @@ export default function ConsultationDialysisTab({
   consultationId,
 }: ConsultationTabProps) {
   return (
-    <div>
-      <PageTitle title="Dialysis Plots" hideBack={true} breadcrumbs={false} />
+    <PageTitle title="Dialysis Plots" hideBack={true} breadcrumbs={false}>
       <DialysisPlots consultationId={consultationId}></DialysisPlots>
-    </div>
+    </PageTitle>
   );
 }
