+++ src/Components/Facility/ConsultationNeurologicalMonitoringTab.tsx
@@ -7,17 +7,16 @@ export default function ConsultationNeurologicalMonitoringTab({
   consultationId,
 }: ConsultationTabProps) {
   return (
-    <div>
-      <PageTitle
-        title="Neurological Monitoring"
-        hideBack={true}
-        breadcrumbs={false}
-      />
+    <PageTitle
+      title="Neurological Monitoring"
+      hideBack={true}
+      breadcrumbs={false}
+    >
       <NeurologicalTable
         facilityId={facilityId}
         patientId={patientId}
         consultationId={consultationId}
       ></NeurologicalTable>
-    </div>
+    </PageTitle>
   );
 }
